fixed misc typos and errors, InternalRobot ready for review

+++ engine/src/main/battlecode/common/GameConstants.java
@@ -61,13 +61,13 @@ public class GameConstants {
     // *********************************
 
     /** A prototype building's starting health, as a multiplier of max health. */
-    public static final float PROTOTYPE_STARTING_HEALTH_MULTIPLIER = 0.1;
+    public static final float PROTOTYPE_STARTING_HEALTH_MULTIPLIER = 0.1f;
 
     /** The maximum level a building can be. */
     public static final int MAX_LEVEL = 3;
 
     /** The amount damage is reduced by every ricochet. */
-    public static final float RICOCHET_DAMAGE_MULTIPLIER = 0.8;
+    public static final float RICOCHET_DAMAGE_MULTIPLIER = 0.8f;
 
     /**
      * The amount of healing an archon does, based on level.
@@ -75,7 +75,7 @@ public class GameConstants {
      * @param level the level of the archon
      * @return the healing amount
      */
-    public static float getArchonHealing(int level) {
+    public static int getArchonHealing(int level) {
         return 8 * level; // TODO: update this. maybe move to RobotType? so both archons and builders have this function
     }
